requirements.txt
aiogram
playwright
beautifulsoup4
aiohttp
python-dotenv
aiosqlite
Pillow